refactor: migrate all imports to difficulty-styles.ts, remove difficulty-colors.ts, fold JsSectionPage inline map
