+++ client/src/lib/difficulty-styles.ts
