+++ client/src/module/student/django/DjangoSectionPage.tsx
@@ -7,7 +7,7 @@ import type { DjangoProgress } from "./data/types";
 import { SEO } from "../../../components/SEO";
 import { canonicalUrl } from "../../../lib/seo.utils";
 import { useAuthStore } from "../../../lib/auth.store";
-import { DIFF_COLOR } from "../../../lib/difficulty-colors";
+import { DIFF_COLOR } from "../../../lib/difficulty-styles";
 
 const FREE_LIMIT = 5;
